requirements.txt
appdirs==1.4.4
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.1
buildozer==1.4.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Cython==0.29.32
decorator==5.1.1
distlib==0.3.6
docutils==0.19
executing==1.2.0
filelock==3.8.2
gTTS==2.3.1
idna==3.4
importlib-metadata==5.2.0
ipython==8.7.0
jedi==0.18.2
Jinja2==3.1.2
Kivy-Garden==0.1.5
kivymd==1.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
pep517==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
playsound==1.3.0
plyer==2.1.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pytoml==0.1.21
requests==2.28.1
sh==1.14.3
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.2
toml==0.10.2
traitlets==5.8.0
urllib3==1.26.13
virtualenv==20.17.1
wcwidth==0.2.5
zipp==3.11.0
